:recycle:  extract utils/logger to package @graphql-markdown/logger
