+++ .github/workflows/test.yml
@@ -20,7 +20,7 @@ jobs:
       fail-fast: false
       matrix:
         package:
-          ["docusaurus", "core", "diff", "printer-legacy", "utils", "helpers"]
+          ["docusaurus", "core", "diff", "printer-legacy", "utils", "helpers", "logger"]
 
     steps:
       - name: Checkout Code Base
