+++ packages/core/package.json
@@ -62,7 +62,8 @@
     "clean": "rm -rf ./dist"
   },
   "dependencies": {
-    "@graphql-markdown/utils": "^1.5.1"
+    "@graphql-markdown/utils": "^1.5.1",
+    "@graphql-markdown/logger": "^1.0.0"
   },
   "peerDependencies": {
     "@graphql-markdown/diff": "^1.0.14",
