+++ packages/core/src/graphql-config.ts
@@ -8,7 +8,7 @@ import type {
   PackageOptionsConfig,
 } from "@graphql-markdown/types";
 
-import { Logger } from "@graphql-markdown/utils";
+import { Logger } from "@graphql-markdown/logger";
 
 const logger = Logger();
