+++ packages/docusaurus/tsconfig.json
@@ -11,5 +11,5 @@
     }
   },
 
-  "references": [{ "path": "../utils" }, { "path": "../core" }]
+  "references": [{ "path": "../utils" }, { "path": "../core" }, { "path": "../logger" }]
 }
