+++ packages/logger/README.md
@@ -1,0 +1,5 @@
+# @graphql-markdown/logger
+
+[![Latest Version](https://img.shields.io/npm/v/@graphql-markdown/logger?style=flat-square)](https://www.npmjs.com/package/@graphql-markdown/logger)
+[![GitHub License](https://img.shields.io/github/license/graphql-markdown/graphql-markdown?style=flat-square)](https://raw.githubusercontent.com/graphql-markdown/graphql-markdown/main/LICENSE)
+
