+++ packages/logger/package.json
@@ -1,0 +1,63 @@
+{
+  "name": "@graphql-markdown/logger",
+  "homepage": "https://graphql-markdown.github.io",
+  "description": "GraphQL-Markdown logger.",
+  "bugs": {
+    "url": "https://github.com/graphql-markdown/graphql-markdown/issues"
+  },
+  "version": "1.0.0",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "graphql-markdown/graphql-markdown",
+    "directory": "packages/logger"
+  },
+  "author": {
+    "name": "Gregory Heitz"
+  },
+  "main": "dist/logger.js",
+  "types": "dist/logger.d.ts",
+  "exports": {
+    ".": {
+      "require": {
+        "types": "./dist/logger.d.ts",
+        "default": "./dist/logger.js"
+      },
+      "import": {
+        "types": "./dist/logger.d.ts",
+        "default": "./dist/logger.js"
+      },
+      "default": {
+        "types": "./dist/logger.d.ts",
+        "default": "./dist/logger.js"
+      }
+    },
+    "./package.json": "./package.json"
+  },
+  "scripts": {
+    "prepack": "pinst --disable",
+    "postpack": "pinst --enable",
+    "stryker": "stryker run",
+    "clean": "rm -rf ./dist",
+    "docs": "typedoc"
+  },
+  "devDependencies": {
+    "@graphql-markdown/types": "^1.0.0"
+  },
+  "directories": {
+    "test": "tests"
+  },
+  "engines": {
+    "node": ">=16.14"
+  },
+  "publishConfig": {
+    "directory": "dist",
+    "access": "public"
+  },
+  "typescript": {
+    "definition": "dist/index.d.ts"
+  },
+  "files": [
+    "dist"
+  ]
+}
