+++ packages/logger/stryker.conf.js
@@ -1,0 +1,6 @@
+// @ts-check
+/**
+ * @type {import('@stryker-mutator/api/core').PartialStrykerOptions}
+ */
+
+module.exports = require("../../config/stryker.conf");
