+++ packages/logger/tests/__data__/dummy_logger.js
