+++ packages/logger/typedoc.json
@@ -1,0 +1,10 @@
+{
+  "extends": ["../../typedoc.base.json"],
+  "entryPoints": ["./src"],
+  "entryPointStrategy": "expand",
+  "out": "./docs",
+  "skipIndexPage": false,
+  "entryFileName": "logger.md",
+  "indexPageTitle": "@graphql-markdown/logger",
+  "exclude": []
+}
