+++ packages/utils/docs/graphql_relation.md
@@ -8,7 +8,7 @@ Library supporting `relatedTypeSection` for displaying relations between GraphQL
 
 ## Functions
 
-### getRelationOfField()
+### getRelationOfField
 
 ```ts
 getRelationOfField(type, schemaMap): Partial< Record< SchemaEntity, RelationOfField[] > >
@@ -16,16 +16,12 @@ getRelationOfField(type, schemaMap): Partial< Record< SchemaEntity, RelationOfFi
 
 Returns a map of fields and arguments where the GraphQL schema type matches the type.
 
-#### See
-
-mapRelationOf
-
 #### Parameters
 
-| Parameter   | Type                     | Description                              |
-| :---------- | :----------------------- | :--------------------------------------- |
-| `type`      | `unknown`                | the GraphQL schema type being processed. |
-| `schemaMap` | `Maybe`\< `SchemaMap` \> | a GraphQL schema map (see getSchemaMap). |
+| Parameter | Type | Description |
+| :------ | :------ | :------ |
+| `type` | `unknown` | the GraphQL schema type being processed. |
+| `schemaMap` | `Maybe`\< `SchemaMap` \> | a GraphQL schema map (see [getSchemaMap](graphql_introspection.md#getschemamap)). |
 
 #### Returns
 
@@ -33,13 +29,17 @@ mapRelationOf
 
 a record map of fields and arguments relations.
 
-#### Source
+#### See
 
-[packages/types/src/utils.d.ts:106](https://github.com/graphql-markdown/graphql-markdown/blob/main/packages/types/src/utils.d.ts#L106)
+mapRelationOf
 
----
+#### Defined In
 
-### getRelationOfImplementation()
+[packages/types/src/utils.d.ts:107](https://github.com/graphql-markdown/graphql-markdown/blob/466abea6/packages/types/src/utils.d.ts#L107)
+
+***
+
+### getRelationOfImplementation
 
 ```ts
 getRelationOfImplementation(type, schemaMap): Partial< Record< SchemaEntity, RelationOfImplementation[] > >
@@ -47,16 +47,12 @@ getRelationOfImplementation(type, schemaMap): Partial< Record< SchemaEntity, Rel
 
 Returns a map of types (unions or interfaces) where the GraphQL schema type is implemented.
 
-#### See
-
-mapRelationOf
-
 #### Parameters
 
-| Parameter   | Type                     | Description                              |
-| :---------- | :----------------------- | :--------------------------------------- |
-| `type`      | `unknown`                | the GraphQL schema type being processed. |
-| `schemaMap` | `Maybe`\< `SchemaMap` \> | a GraphQL schema map (see getSchemaMap). |
+| Parameter | Type | Description |
+| :------ | :------ | :------ |
+| `type` | `unknown` | the GraphQL schema type being processed. |
+| `schemaMap` | `Maybe`\< `SchemaMap` \> | a GraphQL schema map (see [getSchemaMap](graphql_introspection.md#getschemamap)). |
 
 #### Returns
 
@@ -64,13 +60,17 @@ mapRelationOf
 
 a record map of unions or interfaces relations.
 
-#### Source
+#### See
 
-[packages/types/src/utils.d.ts:106](https://github.com/graphql-markdown/graphql-markdown/blob/main/packages/types/src/utils.d.ts#L106)
+mapRelationOf
 
----
+#### Defined In
 
-### getRelationOfInterface()
+[packages/types/src/utils.d.ts:107](https://github.com/graphql-markdown/graphql-markdown/blob/466abea6/packages/types/src/utils.d.ts#L107)
+
+***
+
+### getRelationOfInterface
 
 ```ts
 getRelationOfInterface(type, schemaMap): Partial< Record< SchemaEntity, RelationOfInterface[] > >
@@ -78,16 +78,12 @@ getRelationOfInterface(type, schemaMap): Partial< Record< SchemaEntity, Relation
 
 Returns a map of interfaces where the GraphQL schema type is extended.
 
-#### See
-
-mapRelationOf
-
 #### Parameters
 
-| Parameter   | Type                     | Description                              |
-| :---------- | :----------------------- | :--------------------------------------- |
-| `type`      | `unknown`                | the GraphQL schema type being processed. |
-| `schemaMap` | `Maybe`\< `SchemaMap` \> | a GraphQL schema map (see getSchemaMap). |
+| Parameter | Type | Description |
+| :------ | :------ | :------ |
+| `type` | `unknown` | the GraphQL schema type being processed. |
+| `schemaMap` | `Maybe`\< `SchemaMap` \> | a GraphQL schema map (see [getSchemaMap](graphql_introspection.md#getschemamap)). |
 
 #### Returns
 
@@ -95,13 +91,17 @@ mapRelationOf
 
 a record map of interfaces relations.
 
-#### Source
+#### See
 
-[packages/types/src/utils.d.ts:106](https://github.com/graphql-markdown/graphql-markdown/blob/main/packages/types/src/utils.d.ts#L106)
+mapRelationOf
 
----
+#### Defined In
 
-### getRelationOfReturn()
+[packages/types/src/utils.d.ts:107](https://github.com/graphql-markdown/graphql-markdown/blob/466abea6/packages/types/src/utils.d.ts#L107)
+
+***
+
+### getRelationOfReturn
 
 ```ts
 getRelationOfReturn(type, schemaMap): Partial< Record< SchemaEntity, GraphQLOperationType[] > >
@@ -109,16 +109,12 @@ getRelationOfReturn(type, schemaMap): Partial< Record< SchemaEntity, GraphQLOper
 
 Returns a map of operations (queries, mutations, subscriptions) where the GraphQL schema type is the return type.
 
-#### See
-
-mapRelationOf
-
 #### Parameters
 
-| Parameter   | Type                     | Description                              |
-| :---------- | :----------------------- | :--------------------------------------- |
-| `type`      | `unknown`                | the GraphQL schema type being processed. |
-| `schemaMap` | `Maybe`\< `SchemaMap` \> | a GraphQL schema map (see getSchemaMap). |
+| Parameter | Type | Description |
+| :------ | :------ | :------ |
+| `type` | `unknown` | the GraphQL schema type being processed. |
+| `schemaMap` | `Maybe`\< `SchemaMap` \> | a GraphQL schema map (see [getSchemaMap](graphql_introspection.md#getschemamap)). |
 
 #### Returns
 
@@ -126,13 +122,17 @@ mapRelationOf
 
 a record map of operations relations.
 
-#### Source
+#### See
 
-[packages/types/src/utils.d.ts:106](https://github.com/graphql-markdown/graphql-markdown/blob/main/packages/types/src/utils.d.ts#L106)
+mapRelationOf
 
----
+#### Defined In
 
-### getRelationOfUnion()
+[packages/types/src/utils.d.ts:107](https://github.com/graphql-markdown/graphql-markdown/blob/466abea6/packages/types/src/utils.d.ts#L107)
+
+***
+
+### getRelationOfUnion
 
 ```ts
 getRelationOfUnion(type, schemaMap): Partial< Record< SchemaEntity, GraphQLUnionType[] > >
@@ -140,16 +140,12 @@ getRelationOfUnion(type, schemaMap): Partial< Record< SchemaEntity, GraphQLUnion
 
 Returns a map of unions where the GraphQL schema type is part of it.
 
-#### See
-
-mapRelationOf
-
 #### Parameters
 
-| Parameter   | Type                     | Description                              |
-| :---------- | :----------------------- | :--------------------------------------- |
-| `type`      | `unknown`                | the GraphQL schema type being processed. |
-| `schemaMap` | `Maybe`\< `SchemaMap` \> | a GraphQL schema map (see getSchemaMap). |
+| Parameter | Type | Description |
+| :------ | :------ | :------ |
+| `type` | `unknown` | the GraphQL schema type being processed. |
+| `schemaMap` | `Maybe`\< `SchemaMap` \> | a GraphQL schema map (see [getSchemaMap](graphql_introspection.md#getschemamap)). |
 
 #### Returns
 
@@ -157,6 +153,10 @@ mapRelationOf
 
 a record map of unions relations.
 
-#### Source
+#### See
 
-[packages/types/src/utils.d.ts:106](https://github.com/graphql-markdown/graphql-markdown/blob/main/packages/types/src/utils.d.ts#L106)
+mapRelationOf
+
+#### Defined In
+
+[packages/types/src/utils.d.ts:107](https://github.com/graphql-markdown/graphql-markdown/blob/466abea6/packages/types/src/utils.d.ts#L107)
