+++ packages/utils/docs/object.md
@@ -4,7 +4,7 @@ Internal library of helpers for manipulating objects.
 
 ## Functions
 
-### isEmpty()
+### isEmpty
 
 ```ts
 isEmpty(obj): boolean
@@ -12,10 +12,22 @@ isEmpty(obj): boolean
 
 Check if an object contains key/value records.
 
+#### Parameters
+
+| Parameter | Type | Description |
+| :------ | :------ | :------ |
+| `obj` | `unknown` | the key/value record object. |
+
+#### Returns
+
+`boolean`
+
+`false` if the object is a valid k/v set of records, else `true`.
+
 #### Example
 
 ```js
-import { isEmpty } from "@graphql-markdown/utils/object";
+import { isEmpty } from '@graphql-markdown/utils/object';
 
 const obj = {
   bool: true,
@@ -30,18 +42,6 @@ isEmpty(obj); // Returns false
 isEmpty({}); // Returns true
 ```
 
-#### Parameters
+#### Defined In
 
-| Parameter | Type      | Description                  |
-| :-------- | :-------- | :--------------------------- |
-| `obj`     | `unknown` | the key/value record object. |
-
-#### Returns
-
-`boolean`
-
-`false` if the object is a valid k/v set of records, else `true`.
-
-#### Source
-
-[packages/utils/src/object.ts:34](https://github.com/graphql-markdown/graphql-markdown/blob/main/packages/utils/src/object.ts#L34)
+[packages/utils/src/object.ts:34](https://github.com/graphql-markdown/graphql-markdown/blob/466abea6/packages/utils/src/object.ts#L34)
