+++ packages/utils/docs/prettier.md
@@ -4,13 +4,26 @@ Internal library for prettifying files using `prettier`.
 
 ## Functions
 
-### prettify()
+### prettify
 
 ```ts
 prettify(content, parser): Promise< string | undefined >
 ```
 
 Prettify a string using [prettier.format](https://prettier.io/docs/en/api#prettierformatsource-options).
+
+#### Parameters
+
+| Parameter | Type | Description |
+| :------ | :------ | :------ |
+| `content` | `string` | the string to be prettified. |
+| `parser` | `string` | the `prettier` parser to use. |
+
+#### Returns
+
+`Promise`\< `string` \| `undefined` \>
+
+a prettified string, or undefined if an error occurred.
 
 #### Remarks
 
@@ -20,32 +33,31 @@ This function logs a warning message on error.
 
 https://prettier.io/docs/en/options#parser for the list of parsers.
 
-#### Parameters
+#### Defined In
 
-| Parameter | Type     | Description                   |
-| :-------- | :------- | :---------------------------- |
-| `content` | `string` | the string to be prettified.  |
-| `parser`  | `string` | the `prettier` parser to use. |
+[packages/utils/src/prettier.ts:27](https://github.com/graphql-markdown/graphql-markdown/blob/466abea6/packages/utils/src/prettier.ts#L27)
 
-#### Returns
+***
 
-`Promise`\< `string` \| `undefined` \>
-
-a prettified string, or undefined if an error occurred.
-
-#### Source
-
-[packages/utils/src/prettier.ts:29](https://github.com/graphql-markdown/graphql-markdown/blob/main/packages/utils/src/prettier.ts#L29)
-
----
-
-### prettifyJavascript()
+### prettifyJavascript
 
 ```ts
 prettifyJavascript(content): Promise< string | undefined >
 ```
 
 Prettify a Javascript string using [prettify](prettier.md#prettify) and `babel` parser.
+
+#### Parameters
+
+| Parameter | Type | Description |
+| :------ | :------ | :------ |
+| `content` | `string` | the string to be prettified. |
+
+#### Returns
+
+`Promise`\< `string` \| `undefined` \>
+
+a prettified string, or undefined if an error occurred.
 
 #### Remarks
 
@@ -55,10 +67,24 @@ Same as `prettify(content, "babel")`.
 
 [prettify](prettier.md#prettify)
 
+#### Defined In
+
+[packages/utils/src/prettier.ts:80](https://github.com/graphql-markdown/graphql-markdown/blob/466abea6/packages/utils/src/prettier.ts#L80)
+
+***
+
+### prettifyMarkdown
+
+```ts
+prettifyMarkdown(content): Promise< string | undefined >
+```
+
+Prettify a Markdown string using [prettify](prettier.md#prettify) and `markdown` parser.
+
 #### Parameters
 
-| Parameter | Type     | Description                  |
-| :-------- | :------- | :--------------------------- |
+| Parameter | Type | Description |
+| :------ | :------ | :------ |
 | `content` | `string` | the string to be prettified. |
 
 #### Returns
@@ -66,20 +92,6 @@ Same as `prettify(content, "babel")`.
 `Promise`\< `string` \| `undefined` \>
 
 a prettified string, or undefined if an error occurred.
-
-#### Source
-
-[packages/utils/src/prettier.ts:78](https://github.com/graphql-markdown/graphql-markdown/blob/main/packages/utils/src/prettier.ts#L78)
-
----
-
-### prettifyMarkdown()
-
-```ts
-prettifyMarkdown(content): Promise< string | undefined >
-```
-
-Prettify a Markdown string using [prettify](prettier.md#prettify) and `markdown` parser.
 
 #### Remarks
 
@@ -89,18 +101,6 @@ Same as `prettify(content, "markdown")`.
 
 [prettify](prettier.md#prettify)
 
-#### Parameters
+#### Defined In
 
-| Parameter | Type     | Description                  |
-| :-------- | :------- | :--------------------------- |
-| `content` | `string` | the string to be prettified. |
-
-#### Returns
-
-`Promise`\< `string` \| `undefined` \>
-
-a prettified string, or undefined if an error occurred.
-
-#### Source
-
-[packages/utils/src/prettier.ts:57](https://github.com/graphql-markdown/graphql-markdown/blob/main/packages/utils/src/prettier.ts#L57)
+[packages/utils/src/prettier.ts:59](https://github.com/graphql-markdown/graphql-markdown/blob/466abea6/packages/utils/src/prettier.ts#L59)
