+++ tsconfig.json
@@ -2,11 +2,12 @@
   "files": [],
 
   "references": [
-    { "path": "packages/utils" },
     { "path": "packages/core" },
     { "path": "packages/diff" },
+    { "path": "packages/docusaurus" },
     { "path": "packages/helpers" },
+    { "path": "packages/logger" },
     { "path": "packages/printer-legacy" },
-    { "path": "packages/docusaurus" }
+    { "path": "packages/utils" }
   ]
 }
